Added just a role field to reducer
So as not to break anything in our app, this seems like the most reasonable thing. I went to trying and pull thie role from the new user object in the DashWrapper component and ran into all kinds of croazy errors. So, I determined this was the best option. Let me know what you think? I'll submit the PR and we can go from there.